requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
schedule==1.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
tavily-python==0.3.0
markdown2==2.4.10
openai==1.35.10